+++ src/library/callr/R/result.R
@@ -6,7 +6,7 @@
 #' argument. So we need to check for the error file to decide
 #' if an error happened.
 #'
-#' @param out List of the output object from [run()] and
+#' @param output List of the output object from [run()] and
 #'   the name of the result file to read. For the error file,
 #'   `.error` is appended to this.
 #' @param options The context, including all parameters.
@@ -39,7 +39,7 @@ get_result <- function(output, options) {
   }
 
   ## No error file? Then probably all is well, return the output
-  ## If this is currupt, then the R process has crashed
+  ## If this is corrupt, then the R process has crashed
   ## This cannot happen from R 3.5.0, because that version only writes
   ## out the output file if no error or crash has happened.
   ## (Older R versions write a corrupt RDS file in this case.)