+++ src/library/callr/NEWS.md
@@ -1,3 +1,16 @@
+# callr 3.7.5
+
+* No changes.
+
+# callr 3.7.4
+
+* The `r_session$get_running_time()` method now returns the correct
+  values, as documented (#241, @djnavarro).
+
+* callr now uses fully qualified function calls in the subprocess to
+  avoid interference with functions defined in the global environment.
+  I.e. `base::stderr()` instead of `stderr()`. Closes #246.
+
 # callr 3.7.3
 
 * Errors from callr now include the standard output (in `$stdout`) and